requirements.txt
PyQt6>=6.4.0
pandas>=1.5.0
httpx>=0.23.0
requests>=2.28.1
loguru>=0.6.0
openpyxl>=3.0.10
pillow>=9.2.0
